WIP for supporting codesandbox.io and stackblitz.com
The codesandbox.io part works but currently requires an external
script in `/3rdparty/codesandbox-getparameters.js`. Ideally that
code would be inside live-editor's repo or downloaded from the net.
It's not used unless the user clicks 'codesandbox'

The stackblitz stuff is 90% done? The files make it to stackblitz
but they don't run and I got tired of trying to figure out why.

The UI would still need to be updated to have buttons for these
services