Don't focus if not visible.
When an error happens we move the editor's cursor to the error
and call editor.focus() but that ends up scrolling the parent
page to the error. This is problematic on examples that show
an error since the page jumps immediately to the error.

So, use an IntersectionObserver to see if we're visible.
If not, don't call editor.focus()